This code has been assembled to test the concept of contract free location sensing using magnetometers. Magnetic field vector data for two sensors is pulled from the serial port of an Arduino board. Time is allowed for the sensors to be calibrated within the hardware. The input data is manipulated for easy management. Key assumpions are made to enable calculations, which include an empirically derived formula for distance. The main assumptions are that the magnet is always on a plane and that the field is emitted symetrically in all directions and as a point dipole of negligible volume. Absolute distance to sensor 0 and sensor 1, the approximated x and y coordinates of the magnet, and the calculated angle between a line connecting the two sensors and the location of the magnet (theta) are printed to stdout. The approximated values for x and y are also plotted in a real time loop to a scatter plot.
There are presently many limitations to the program. Though the empirical formula for raw distance from the sensors is both rather accurate and precise, the derived values acquired from those distances are not. It is not yet possible to determine on what side of the x-axis the magnet is on without amplifying a gross error. The framerate of the plot is about 2fps. The range for theta does not generally include values below 10 degrees above the x axis. And most importantly, the range of the (main) sensor 0, and so the approximated location of x and y, is limited by the mathematical restrictions on solutions to the triangle formed by the location of the magnet and the two sensors. It is possible that many of these limitations will be mitigated by better design of the sensor circuit, and a stronger, smaller, more homogeneous magnet.

However, for this code, it should be possible to:

1. Increase the range of solvable triangles from the input data, providing for a larger, plottable sensor range.
2. Implement a better derived empirical formula.
3. Increase the framerate of the plot.
4. Enable calibration/reset commands to the sensor.
5. Include approximate location information for a z-axis***

(***Note: Difficult.)
-EDN