Add files via upload
Both addresses are enabled to improve data sampling rates (s.t. half as many MUX switches are necessary).
The right hand rule was enforced on sensor output to reflect a coordinate system definition that makes physical sense, and is consistent with the rest of the algorithm; the sensor's built-in C.S. was essentially thrown out.